pyproject.toml
[tool.poetry]
name = "sorayiaapi"
version = "0.1.0"
description = ""
authors = ["sona-user-369 <donadavak@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
api = "0.0.7"
asgiref = "3.7.2"
certifi = "2023.11.17"
charset-normalizer = "3.3.2"
click = "8.1.7"
django = "4.2.7"
django-rest-framework = "0.1.0"
djangorestframework = "3.14.0"
idna = "3.6"
nose = "1.3.7"
pytz = "2023.3.post1"
requests = "2.31.0"
sqlparse = "0.4.4"
urllib3 = "2.1.0"
dj-database-url = "^2.1.0"
psycopg2-binary = "^2.9.9"
gunicorn = "^21.2.0"
drf-yasg = "^1.21.7"
whitenoise = {extras = ["brotli"], version = "^6.6.0"}
i = "^2.1.0"
django-cors-headers = "^4.3.1"
firebase-admin = "^6.3.0"
langchain = ">=0.0.327,<0.1.0"
python-decouple = "^3.8"
openai = "0.28.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
